chore(rebase): prune old worktrees
this ensures that the prepare call for new events on the same branch does not
fail.

alternative approaches would include to re-use previous worktrees, but then we
need to ensure that we're properly up to date with master. I don't have a
preferrence for either approach, so if this doesn't work reliably I'd switch
approaches